Extract QSCriptCorePlanner
